+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/AutomatedCheckResponse.java
@@ -1,6 +1,5 @@
 package org.grnet.cat.dtos;
 
-import com.fasterxml.jackson.annotation.JacksonAnnotation;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;